Desenvolvimento CRUD

+++ MID-PLATFORM/Controllers/CompaniesController.cs
@@ -20,6 +20,7 @@ namespace MID_PLATFORM.Controllers
             _context = context;
         }
 
+        //READ
         // GET: api/Companies
         [HttpGet]
         public async Task<ActionResult<IEnumerable<Company>>> GetCompanies()
@@ -31,6 +32,7 @@ namespace MID_PLATFORM.Controllers
             return await _context.Companies.ToListAsync();
         }
 
+        //READ
         // GET: api/Companies/5
         [HttpGet("{id}")]
         public async Task<ActionResult<Company>> GetCompany(int id)
@@ -49,6 +51,7 @@ namespace MID_PLATFORM.Controllers
             return company;
         }
 
+        //UPDATE
         // PUT: api/Companies/5
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPut("{id}")]
@@ -59,7 +62,18 @@ namespace MID_PLATFORM.Controllers
                 return BadRequest();
             }
 
-            _context.Entry(company).State = EntityState.Modified;
+            //_context.Entry(company).State = EntityState.Modified;
+
+            Company modifiedCompany = _context.Companies.FirstOrDefault(c => c.CompanyId == id);
+            if (modifiedCompany == null)
+            {
+                return NotFound();
+            }
+
+            modifiedCompany.Code = company.Code;
+            modifiedCompany.Name = company.Name;
+            modifiedCompany.Country = company.Country;
+            modifiedCompany.Active = company.Active;
 
             try
             {
@@ -80,40 +94,80 @@ namespace MID_PLATFORM.Controllers
             return NoContent();
         }
 
+        //CREATE
         // POST: api/Companies
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPost]
         public async Task<ActionResult<Company>> PostCompany([FromBody]Company company)
         {
-
             if (_context.Companies == null)
             {
                 return Problem("Entity set 'MIDPlatformContext.Companies'  is null.");
             }
             _context.Companies.Add(company);
-            await _context.SaveChangesAsync();
+            try
+            {
+                await _context.SaveChangesAsync();
+            }
+            catch(Exception e)
+            {
+                return Problem(e.InnerException.ToString(),null,null,e.Message);
+            }
 
             return CreatedAtAction("GetCompany", new { id = company.CompanyId }, company);
         }
 
+        //DELETE
         // DELETE: api/Companies/5
+        [HttpDelete("{id},{realDelete}")]
         [HttpDelete("{id}")]
-        public async Task<IActionResult> DeleteCompany(int id)
+        public async Task<IActionResult> DeleteCompany(int id, bool disable = false)
         {
             if (_context.Companies == null)
             {
                 return NotFound();
             }
+
             var company = await _context.Companies.FindAsync(id);
             if (company == null)
             {
                 return NotFound();
             }
 
-            _context.Companies.Remove(company);
-            await _context.SaveChangesAsync();
+            if (disable)//se for true desativa, se for false apaga
+            {
+                company.Active = false;
+                _context.Companies.Update(company);
+            }
+            else
+            {
+                _context.Companies.Remove(company);
+            }
 
-            return NoContent();
+            try
+            {
+                await _context.SaveChangesAsync();
+            }
+            catch (DbUpdateException ex)
+            {
+                try
+                {
+                    company.Active = false;
+                    _context.Companies.Update(company);
+
+                    return Ok(ex.InnerException);
+                }
+                catch (Exception e)
+                {
+                    return Problem(e.InnerException.ToString(), null, null, e.Message);
+                }
+            }
+            catch (Exception e)
+            {
+                return Problem(e.InnerException.ToString(), null, null, e.Message);
+            }
+
+            return Ok();
         }
 
         private bool CompanyExists(int id)
